rcar-gen3: kernel-module-uvcs: add UVCS driver
UVCS driver is required to run Multimedia functions.

UVCS Driver is the function that provides access interface to the video hardware,
scheduler, arbiter, and interrupt handler.
To access to this driver or to call each function in this driver is not required
except for installation of this driver. These operations are implemented to
related Renesas proprietary software.

Signed-off-by: Phong Man Tran <phong.tran.wh@rvc.renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>