rcar-gen3: core-image-weston: fix the wrong wordsize issue
This patch fixes the wrong worksize issue when building the
lib32-core-image-weston and lib32-core-image-weston-sdk.

The issue occurs due to some 64 bit pkgs (aarch64)
are installed to 32 bit SDK. That makes the override the libc-dev
32 bit headers by libc-dev 64 bit headers.

The 64 bit pkgs are installed to SDK due to the dependecy pkgs.
This patch will remove the packages that lead to install 64 bit
pkgs.

List of them are: weston-init, kernel-devsrc, liberation-fonts.

Signed-off-by: Phong Man Tran <phong.tran.wh@rvc.renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>