rcar-gen3: linuxptp: Fix miss setting problem and Add install config files
This patch fixes the miss setting problem with the ravb driver for
the R-Car Linux BSP, And add install the default configuration files
for running linuxptp.

The detail description of fixes the miss setting problem.

The ravb driver is supported the following the timestamp filter.
- HWTSTAMP_FILTER_ALL
- HWTSTAMP_FILTER_PTP_V2_L2_EVENT (recommended)

If using the HWTSTAMP_FILTER_PTP_V2_L2_EVENT, the ravb driver work
more efficiently than when the HWTSTAMP_FILTER_ALL.
So recommended to use HWTSTAMP_FILTER_PTP_V2_L2_EVENT.
But linuxptp does not use it.

The following patches are fixes this problem.
- 0001-sk-fix-error-check-when-hwts_init.patch
- 0002-sk-replaced-the-priority-of-hwtstamp-filter1-and-fil.patch

Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>