rcar-gen3: avb: Add support for AVB software package
This patch adds AVB:

Linux kernel module for AVB (kernel-module-avb)
Demo and utility applications for AVB (avb-applications)
The Open-AVB project by AVnu (open-avb)

The AVB is disable as default in mmp/local-wayland.conf.
To enable the AVB modify #DISTRO_FEATURES_append = " avb"
-> DISTRO_FEATURES_append = " avb"

[kernel-module-avb]
The kernel-module-avb is Linux kernel modules for AVB software package.

This patch added the following modules.

- kernel-module-avb-streaming: AVB Streaming Driver.
- kernel-module-avb-mse: AVB MSE Driver

Note: This patch added the usb-video-class.cfg in the linux-renesas
files. This config fragment is needed for compile of the AVB MSE
Driver.

[avb-applications]
The avb-applications is Demo and Utility for AVB software package.

This patch added the following applications.

- avbtool: Utility tool of the AVB Streaming Driver
- avb-demoapps: Renesas Ethernet AVB demo applications

[Open-AVB]
The Open-AVB project is to provide components for bulding AVB systems.
This project contains components - drivers, libraries, examples
application and protocol stack daemons - required to build and AVB system.

This patch added the following protocol stack daemons.

- open-avb-gptp: gPTP (generalized Precision Time Protocol) daemon
- open-avb-mrpd: SRP (Stream Reservation Protocol) daemon
- open-avb-maap: MAAP (MAC address acquisition protocol) daemon

Signed-off-by: Masaru Nagai <masaru.nagai.vx@renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>