vspmif-user-module: fix header installation
The do_compile step was copying directly to the sysroot's
/usr/include. Because this was a direct copy it was not put into
the sstate.
The do_install step was installing the same header files into
/usr/local/include. Because includedir was not pointed here, they
were also not put into the sstate.
The issue was then that if you tried to do a build reusing sstate,
the header files from vspmif-user-module would not be present,
and then dependent recipes like vspmif-tp-user-module.

The patch fixes this by removing the direct copy to sysroot,
and setting includedir to /usr/local/include.

Signed-off-by: Frederico Cadete <frederico.cadete@awtce.be>