rcar-gen3: conf: changed GCC ARM option to -mtune=cortex-a57.cortex-a53
In GCC ARM option of mtune, cortex-a57.cortex-a53 can be used for
big.Little system.

This patch adds the tuning file for -mtune=cortex-a57.cortex-a53.
And changed mtune option to cortex-a57.cortex-a53 from cortex-a57.

Even though the current launched CPU is cortex-A57 only, this option
will be the best in the future.

Signed-off-by: Phong Man Tran <phong.tran.wh@rvc.renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>