rcar-gen3: kernel-module-gles: WORKAROUND to build with GCC 8.x
Current GFX source code will get compilation error with GCC 8.x,
because it is defined in makefile to use -Wall to enable all
Warnings check.
-Wcast-function-type, -Wsizeof-pointer-div and -Wstringop-truncation
are new command-line options have been added for the C and C++
compilers in GCC 8.

This commit applies a W/A patch to silence -Wcast-function-type,
-Wsizeof-pointer-div and -Wstringop-truncation errors.

Signed-off-by: Thuy Tran <thuy.tran.xh@renesas.com>
Signed-off-by: Khang Nguyen <khang.nguyen.xw@renesas.com>
Signed-off-by: Takamitsu Honda <takamitsu.honda.pv@renesas.com>