rcar-gen3: omx-module: Improve vp9dec_lib checking function
This commit modifies the vp9dec_lib checking function to disable
vp9dec_lib automatically when not building for M3N, E3 and also warns
the user to build it on M3 v3.0 only.
This also allows to run the checking at the end of recipe parsing,
not have to wait until do_patch finished.

Change-Id: I6329c223a42d10a041407a349bac77927b5f3614
Signed-off-by: Duy Dang <duy.dang.yw@renesas.com>
Signed-off-by: Khang Nguyen <khang.nguyen.xw@renesas.com>
Signed-off-by: Takamitsu Honda <takamitsu.honda.pv@renesas.com>