rcar-gen3: kernel-module-gles: WORKAROUND to build with GCC 9.x
Current GFX source code will get compilation error with GCC 9.x,
because it is defined in makefile to use -Wall to enable all
Warnings check.

-Wcast-function-type, -Wsizeof-pointer-div and -Wstringop-truncation,
-Wmissing-attributes, -Waddress-of-packed-member are new command-line
options have been added for the C and C++ compilers in GCC 9.

This commit applies a W/A patch to silence -Wcast-function-type,
-Wsizeof-pointer-div, -Wstringop-truncation, -Wmissing-attributes
and -Waddress-of-packed-member errors.

Signed-off-by: Khang Nguyen <khang.nguyen.xw@renesas.com>
Signed-off-by: Takamitsu Honda <takamitsu.honda.pv@renesas.com>
Change-Id: Ic14200e68d0c3a666f9ef1906a3e40b9d6a721dc