rcar-gen3: conf: Drop deprecated fix in GFX configuration
In GFX sample configuration, the gstreamer1.0-plugins-bad is set
RDEPENDS on libwayland-egl to fix a warning, this warning does not
occur in new software versions.

In addition, from Wayland 1.15.0 official announcement:
https://lists.freedesktop.org/archives/wayland-devel/2018-April/037767.html

| libwayland-egl is now part of libwayland, and will presumably be
| removed from mesa in the not too distant future.

The RDEPENDS will cause a dependency error due to the above update.
This commit removes the fix to avoid the error.

Change-Id: I1bd01a2b79cee04c34ac5de2dddc03e86558ecf4
Signed-off-by: Khang Nguyen <khang.nguyen.xw@renesas.com>
Signed-off-by: Takamitsu Honda <takamitsu.honda.pv@renesas.com>