rcar-gen3: base-files: Fix missing log and tmp in /var/volatile
Since systemd v240, tmpfiles could not create the files/directories
under a prefix that may be owned by an unprivileged user due to unsafe
transition validation.

In case of users set up a rootfs on NFS server or use PC to prepare
one on a flash device, the ownership is inherited from host PC, but
the /var/volatile is set the owner to root by default, it leads to
an issue that the log and tmp directories under /var/volatile are not
created. The processes use the tmp/log directories to store log and
temporary files will fail.

| [FAILED] Failed to start Login Service.
| [FAILED] Failed to start Network Name Resolution.
| authpriv.err login[3074]: pam_lastlog(login:session): unable to create /var/log/lastlog: No such file or directory

This changes the owner of /var/volatile to uid=1000,gid=1000 as a
workaround for the issue.

Signed-off-by: Khang Nguyen <khang.nguyen.xw@renesas.com>
Signed-off-by: Takamitsu Honda <takamitsu.honda.pv@renesas.com>
Change-Id: Iae54eb2f47cca2d7f57c48d581d075b6c41401b7